Move SliceManagerAPI client to Orchestrator
